+++ intake_esm/bld_collection_utils.py
@@ -3,15 +3,16 @@ import hashlib
 import os
 import re
 import urllib
-from glob import glob
 from pathlib import Path
 from urllib.request import urlopen, urlretrieve
 
+import pandas as pd
 import xarray as xr
 import yaml
 from intake.source.utils import reverse_format
 
 from . import config
+from .storage import _get_hsi_files, _posix_symlink
 
 _default_cache_dir = config.get('database-directory')
 _default_cache_dir = f'{_default_cache_dir}/bld-collection-input'
@@ -281,3 +282,78 @@ def _reverse_filename_format(file_basename, filename_template=None, gridspec_tem
                 f'Failed to parse file: {file_basename} using patterns: {filename_template} and {gridspec_template}'
             )
             return {}
+
+
+def _filter_query_results(query_results, file_basename_column_name):
+    """Filter for entries where file_basename is the same and remove all
+       but the first ``direct_access = True`` row."""
+
+    groups = query_results.groupby(file_basename_column_name)
+
+    gps = []
+    for _, group in groups:
+
+        g = group[group['direct_access']]
+        # File does not exist on resource with high priority
+        if g.empty:
+            gps.append(group)
+
+        else:
+            gps.append(g)
+
+    query_results = pd.concat(gps)
+    return query_results
+
+
+def _ensure_file_access(
+    query_results,
+    file_fullpath_column_name='file_fullpath',
+    file_basename_column_name='file_basename',
+):
+    """Ensure that requested files are available locally.
+
+    Paramters
+    ---------
+    query_results : `pandas.DataFrame`
+        Results of a query.
+
+    Returns
+    -------
+    local_urlpaths : list
+        List of urls to access files in `query_results`.
+    """
+
+    resource_types = {'hsi': _get_hsi_files, 'copy-to-cache': _posix_symlink}
+
+    data_cache_directory = config.get('data-cache-directory')
+
+    os.makedirs(data_cache_directory, exist_ok=True)
+
+    file_remote_local = {k: [] for k in resource_types.keys()}
+
+    query_results = _filter_query_results(query_results, file_basename_column_name)
+
+    local_urlpaths = []
+    for idx, row in query_results.iterrows():
+        if row.direct_access:
+            local_urlpaths.append(row[file_fullpath_column_name])
+
+        else:
+            file_remote = row[file_fullpath_column_name]
+            file_local = os.path.join(data_cache_directory, os.path.basename(file_remote))
+            local_urlpaths.append(file_local)
+
+            if not os.path.exists(file_local):
+                if row.resource_type not in resource_types:
+                    raise ValueError(f'unknown resource type: {row.resource_type}')
+
+                file_remote_local[row.resource_type].append((file_remote, file_local))
+
+    for res_type in resource_types:
+        if file_remote_local[res_type]:
+            print(f'transfering {len(file_remote_local[res_type])} files')
+            resource_types[res_type](file_remote_local[res_type])
+
+    query_results[file_fullpath_column_name] = local_urlpaths
+
+    return query_results
